chore: move poseidonHash to num-theory to avoid delays in writing files
