+++ src/boilerplate/common/number-theory.mjs
@@ -5,6 +5,7 @@
 // We are very grateful for the original work by rsandor
 import config from 'config';
 import pkg from 'general-number';
+import poseidonConstants from './poseidon-constants.mjs';
 
 const { BN128_PRIME_FIELD, BN128_GROUP_ORDER, BABYJUBJUB } = config;
 
@@ -13,6 +14,8 @@ const one = BigInt(1);
 const { JUBJUBE, JUBJUBC, JUBJUBD, JUBJUBA } = BABYJUBJUB;
 const Fp = BN128_GROUP_ORDER; // the prime field used with the curve E(Fp)
 const Fq = JUBJUBE / JUBJUBC;
+const DOMAIN_KEM = BigInt(10);
+const DOMAIN_DEM = BigInt(20);
 
 function addMod(addMe, m) {
   return addMe.reduce((e, acc) => (((e + m) % m) + acc) % m, BigInt(0));
@@ -303,8 +306,71 @@ function decrypt(encryptedMessages, secretKey, encPublicKey) {
   encryptedMessages.forEach((msg, index) => {
     const hash = poseidonHash([key.bigInt, BigInt(DOMAIN_DEM), BigInt(index)]);
     plainText[index] = addMod([BigInt(msg), -hash.bigInt], Fp);
+    while (plainText[index] < 0n) {
+      plainText[index] += Fp;
+    }
   });
   return plainText;
+}
+
+// Implements the Poseidon hash, drawing on the ZoKrates implementation
+// roundsP values referred from circom library
+// https://github.com/iden3/circomlibjs/blob/main/src/poseidon_opt.js
+
+const { C, M, SNARK_SCALAR_FIELD: q } = poseidonConstants;
+
+function ark(state, c, it) {
+  const N = state.length;
+  for (let i = 0; i < N; i++) {
+    state[i] = addMod([state[i], c[it + i]], q);
+  }
+  return state;
+}
+
+function sbox(state, f, p, r) {
+  const N = state.length;
+  state[0] = powerMod(state[0], 5n, q);
+  for (let i = 1; i < N; i++) {
+    state[i] = r < f / 2 || r >= f / 2 + p ? powerMod(state[i], 5n, q) : state[i];
+  }
+  return state;
+}
+
+function mix(state, m) {
+  const N = state.length;
+  const out = new Array(N).fill(0n);
+  for (let i = 0; i < N; i++) {
+    let acc = 0n;
+    for (let j = 0; j < N; j++) {
+      acc = addMod([acc, mulMod([state[j], m[i][j]], q)], q);
+    }
+    out[i] = acc;
+  }
+  return out;
+}
+
+function poseidonHash(_inputs) {
+  if (_inputs.length > 16) throw new Error('To many inputs to Poseidon hash');
+  const inputs = _inputs;
+  const N = inputs.length;
+  const t = N + 1;
+  const roundsP = [56, 57, 56, 60, 60, 63, 64, 63, 60, 66, 60, 65, 70, 60, 64, 68];
+  const f = 8;
+  const p = roundsP[t - 2];
+  const c = C[t - 2];
+  const m = M[t - 2];
+
+  let state = new Array(t).fill(0n);
+  for (let i = 1; i < t; i++) {
+    state[i] = inputs[i - 1];
+  }
+  for (let r = 0; r < f + p; r++) {
+    state = ark(state, c, r * t);
+    state = sbox(state, f, p, r);
+    state = mix(state, m);
+  }
+  // console.log('MATRIX', m);
+  return generalise(state[0]);
 }
 
 // These exports are not unused, but find-unused-exports linter will complain because they are not used
@@ -319,4 +385,5 @@ export {
   compressStarlightKey,
   decompressStarlightKey,
   decrypt,
+  poseidonHash,
 };
